Rename CommandHandler and add register
